fix(#1015): clear the clippy backlog through every masking layer; 23 errors remain, tracked in #1020
Clears the clippy lint backlog that blocked the repo-wide gate (#965), across the layers that were
masking each other: a crate failing its own `-D warnings` lints blocks its dependents from being
linted at all, so each pass revealed errors the previous pass could not see.

What this does NOT do, stated plainly because the branch's own first commit subject claimed
otherwise: **`cargo clippy --workspace --all-targets -D warnings` does not exit 0 after this
change.** 23 errors remain, measured behind an `#[allow]` scaffold that unmasked every layer and
confirmed there is no further one. They are tracked in #1020. The scaffold bounds `--all-targets`
clippy only; clippy does not lint doctests.

Also restores NaN semantics that the `.max().min()` -> `.clamp()` rewrite silently changed
(`src/movement.rs`). The exposed case was the DEFAULT configuration, not an edge case: with the
default `-infinity` underworld, `ref_z = from[2].max(-infinity)` is `-infinity` for a non-finite
input, and `-infinity - (-infinity)` is NaN. A finite underworld is the safe case, because `max`
washes the NaN out. Found by the independent reviewer, whose own round-1 reasoning had the
implication backwards and who corrected it by compiling and running the four-case table rather than
reasoning about it a second time.

Independently reviewed over two rounds. The reviewer rebuilt the 23-site scaffold from scratch rather
than reusing the author's, hit and fixed its own placement bug along the way, and re-derived the test
gate itself: exit 0, 55/55 header/result pairs each individually checked for internal consistency,
2079 passed / 0 failed / 50 ignored / 0 filtered out, 13 doctest targets.

Round 1 refuted a false exhaustiveness claim in the PR body ("zero further errors beyond the 11
already listed"). It was struck at the source and traced to a copy on issue #1015, which was struck
too; a concept-level grep over tracked files and commit messages found no third copy.

Closes #1015.

+++ crates/eqoxide-http/src/lib.rs
@@ -293,7 +293,7 @@ impl PlayerState {
             zone_in_failed: gs.world.zone_in_failed,
             race:       gs.player_race.clone(),
             class:      gs.player_class.clone(),
-            level:      gs.player_level as u32,
+            level:      gs.player_level,
             pos_east:   gs.player_x,
             pos_north:  gs.player_y,
             position_provisional_since: gs.position_provisional_since,
